+- add patches.fixes/linux-post-2.6.3-rc2-20040212 +  bluetooth, watchdog, cpufreq, ia64 updates
suse-commit: bcfa623cb01aafcac0675812dd2b6bcae111c4c4

+++ include/asm-ppc64/vio.h
@@ -97,7 +97,7 @@ static inline struct vio_driver *to_vio_driver(struct device_driver *drv)
 struct vio_dev {
 	struct device_node *archdata;   /* Open Firmware node */
 	void *driver_data;              /* data private to the driver */
-	unsigned long unit_address;	
+	unsigned int unit_address;	
 	struct TceTable *tce_table;     /* vio_map_* uses this */
 	unsigned int irq;
